+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `Builder.close` method has optional argument - count iterations of "closing"
+- `classList` attribute
+
+### Changed
+
+- Error message of raised exception in case of wrong state
+
 ## [0.6.0] - 2023-08-21
 
 ### Added
@@ -68,6 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Wrap all builded elements by provided DOM node
 
 [Unreleased]: https://github.com/vikian050194/fandom/compare/v0.6.0...HEAD
+[0.7.0]: https://github.com/vikian050194/fandom/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/vikian050194/fandom/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/vikian050194/fandom/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/vikian050194/fandom/compare/v0.3.0...v0.4.0